requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0
aiosmtplib>=2.0
ntplib>=0.4.0
pyyaml>=6.0
loguru>=0.7.0
